+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/living/LivingShops.java
@@ -45,10 +45,24 @@ public class LivingShops {
 
 				// TODO Remove this migration again at some point
 				// MC 1.16:
-				// Convert pig-zombie to zombified-piglin (but only if we run on MC 1.16 or above):
+				// Convert 'pig-zombie' to 'zombified-piglin':
 				if (objectTypeId.equals("pig-zombie")) {
 					shopObjectData.set(AbstractShopObject.SHOP_OBJECT_TYPE_ID, "zombified-piglin");
 					Log.warning(logPrefix + "Migrated object type 'pig-zombie' to 'zombified-piglin'.");
+					migrated = true;
+				}
+
+				// MC 1.20.5:
+				// Convert 'mushroom-cow' to 'mooshroom':
+				if (EntityType.MUSHROOM_COW.name().equals("MOOSHROOM") && objectTypeId.equals("mushroom-cow")) {
+					shopObjectData.set(AbstractShopObject.SHOP_OBJECT_TYPE_ID, "mooshroom");
+					Log.warning(logPrefix + "Migrated object type 'mushroom-cow' to 'mooshroom'.");
+					migrated = true;
+				}
+				// Convert 'snowman' to 'snow-golem':
+				if (EntityType.SNOWMAN.name().equals("SNOW_GOLEM") && objectTypeId.equals("snowman")) {
+					shopObjectData.set(AbstractShopObject.SHOP_OBJECT_TYPE_ID, "snow-golem");
+					Log.warning(logPrefix + "Migrated object type 'snowman' to 'snow-golem'.");
 					migrated = true;
 				}