+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/living/SKLivingShopObjectTypes.java
@@ -80,7 +80,7 @@ import com.nisovin.shopkeepers.util.java.Validate;
  * <li>IRON_GOLEM: okay
  * <li>MAGMA_CUBE: okay, would usually spawn with random size, weird behavior in water (no longer
  * the case, maybe due to using NoAI)
- * <li>MUSHROOM_COW: okay
+ * <li>MUSHROOM_COW: okay, renamed to MOOSHROOM in Spigot 1.20.5
  * <li>OCELOT: okay
  * <li>PIG: okay
  * <li>PIG_ZOMBIE: okay; replaced by ZOMBIFIED_PIGLIN in MC 1.16
@@ -89,7 +89,7 @@ import com.nisovin.shopkeepers.util.java.Validate;
  * NoAI
  * <li>SKELETON: okay
  * <li>SLIME: okay, would usually spawn with random size
- * <li>SNOWMAN: okay
+ * <li>SNOWMAN: okay, renamed to SNOW_GOLEM in Spigot 1.20.5
  * <li>SPIDER: okay
  * <li>SQUID: seems okay, slightly weird movement in water
  * <li>WITCH: okay
@@ -185,8 +185,11 @@ public final class SKLivingShopObjectTypes implements LivingShopObjectTypes {
 		Map<@NonNull EntityType, @NonNull List<? extends @NonNull String>> aliases = new HashMap<>();
 		aliases.put(EntityType.MUSHROOM_COW, prepareAliases(Arrays.asList(
 				"mooshroom",
-				"mooshroom-cow",
-				"mushroom"
+				"mushroom-cow"
+		)));
+		aliases.put(EntityType.SNOWMAN, prepareAliases(Arrays.asList(
+				"snow-golem",
+				"snowman"
 		)));
 		ALIASES = Collections.unmodifiableMap(aliases);
 	}